feat: unwrap permission namespace to flat exports + barrel
